Estimator

SLAM.Estimator — Type

Estimator is responsible for performing Bundle-Adjustment procedure and map filtering that removes low-information Keyframes from the map.

Arguments:

map_manager::MapManager: Map manager with all the map information.

params::Params: Parameters of the system.

frame_queue::Vector{Frame}: Queue of the Frames to process.

new_kf_available::Bool: Whether or not new Keyframe in the queue is available.

exit_required::Bool: Whether the estimator should end its work and exit.

queue_lock:ReentrantLock: Lock for the queue.

SLAM.run! — Method

run!(estimator::Estimator)

Main routine that starts Estimator.

SLAM.add_new_kf! — Function

add_new_kf!(estimator::Estimator, frame::Frame)

Add frame to the Estimator queue to be processed.

SLAM.get_new_kf! — Function

get_new_kf!(estimator::Estimator)

Get frame from the queue if it is available.

SLAM.local_bundle_adjustment! — Function

local_bundle_adjustment!(estimator::Estimator, new_frame::Frame)

Perform Bundle-Adjustment on the new frame and its covisibility graph.

Minimize error function over all KeyFrame's extrinsic parameters in the covisibility graph and their corresponding MapPoint's positions. Afterwards, update these parameters.

SLAM.map_filtering! — Function

map_filtering!(estimator::Estimator, new_keyframe::Frame)

Filter out KeyFrames that share too many MapPoints with other KeyFrames in the covisibility graph. Since they are not informative.

SLAM.reset! — Method

reset!(estimator::Estimator)

Reset Estimator.