IO

SLAM.Visualizer — Type

Visualizer for the SLAM.

Can be used to play in "live" mode as the SLAM is working or to replay from ReplaySaver.

Usage:

To use it in "live" preview mode, pass it as a keyword argument, when creating SlamManager:

visualizer = Visualizer(;resolution=(900, 600), image_resolution=(370, 1226))
SlamManager(params, camera; visualizer)

SLAM.set_frame_wc! — Method

set_frame_wc!(v::Visualizer, frame_id, wc)

Add new frame wc to the visualizer queue. This is used when other threads are updating the visualizer.

The queue is processed in process_frame_wc! method.

SLAM.process_frame_wc! — Function

process_frame_wc!(v::Visualizer)

Process pose queue. This updates the plot immediately.

SLAM.set_image! — Function

set_image!(v::Visualizer, image)

Update image in the visualizer. It should have the same dimensions as the original.

SLAM.set_position! — Function

set_position!(v::Visualizer, position)

Add new position to the camera positions. This updates the plot immediately.

SLAM.ReplaySaver — Type

ReplaySaver()

ReplaySaver is used to save/load SLAM result for the visualizer. If added to the SlamManager instead of the visualizer, it accumulates poses needed for the visualization/replay and can be later used to replay the results.

Usage:

To use it, pass it to visualizer keyword argument, when creating SlamManager. Do some work, and in the end, save it to the directory.

saver = ReplaySaver()
SlamManager(params, camera; visualizer=saver)
# ...run slam...
SLAM.save(saver, "save_dir")

Then you can load it from the save directory and use it to replay results.

SLAM.load!(saver, "save_dir")

SLAM.set_frame_wc! — Method

set_frame_wc!(saver::ReplaySaver, frame_id, wc)

Add new pose to store.

SLAM.save — Function

save(saver::ReplaySaver, save_dir)

Save results into the given directory.

SLAM.load! — Function

load!(saver::ReplaySaver, save_dir)

Load results from a given directory.