Slam Manager

SLAM.SlamManager — Type

SlamManager(
    params::Params, camera::Camera;
    right_camera::Union{Nothing, Camera} = nothing,
)

Slam Manager that is the highest level component in the system. It is responsible for sending new frames to the other components and for processing their outputs.

Note, that upon creating, SlamManager launches Mapper in the separate thread.

Arguments

params::Params: Parameters of the system.

image_queue::Vector{Matrix{Gray{Float64}}}: Queue of the images to be processed.

right_image_queue::Vector{Matrix{Gray{Float64}}}: In case of stereo mode, queue of images for the right camera. It should be in sync with image_queue which in this case is for the left camera.

time_queue::Vector{Float64}: Queue of timestamps for each of the frame. The timestamps are used in the motion model to predict next pose for the frame, before it is actually computed.

current_frame::Frame: Current frame that is processed. It is shared among all other components in the system.

frame_id::Int64: Id of the current frame.

front_end::FrontEnd: Front-End component that is used for tracking.

map_manager::MapManager: Map manager for the managment of keyframes & mappoints.

mapper::Mapper: Mapper that is used for triangulation of keypoints. It is launched in the constructor as a separate thread.

extractor::Extractor: Used in extraction of keypoints from the frames.

exit_required::Bool: Set it to true to stop SlamManager once it is launched.

SLAM.run! — Method

run!(sm::SlamManager)

Main routine for the SlamManager. It runs until exit_required variable is set to true. After that, it will end its work, wait for other threads and finish.

Once there is a frame in the queue, it will get it and first send it to the FrontEnd for tracking. If FrontEnd requires a new Keyframe, then it will also send it to the mapper thread for Keyframe creation and triangulation.

SLAM.add_image! — Function

add_image!(sm::SlamManager, image, time)

Add monocular image and its timestamp to the queue.

SLAM.add_stereo_image! — Function

add_stereo_image!(sm::SlamManager, image, right_image, time)

Add stereo image and its timestamp to the queue.

SLAM.get_queue_size — Function

get_queue_size(sm::SlamManager)

Get size of the queue of images to be processed.

SLAM.reset! — Method

reset!(sm::SlamManager)

Reset slam manager, front-end and map_manager.