API Reference

SLAM.Extractor — Type

Feature extractor.

Performs Shi-Tomasi feature extraction algorithm with support for masking-out regions in which to avoid feature detection.

SLAM.Frame — Type

Frame that encapsulates information of a camera in space at a certain moment.

Parameters:

id::Int64: Id of the Frame.

kfid::Int64: Id of the Frame in the global map, contained in MapManager. Frames that are in this map are called "Key-Frames".

time::Float64: Time of the frame at which it was taken on the camera.

cw::SMatrix{4, 4, Float64, 16}: Transformation matrix [R|t] that transforms from world to camera space.

wc::SMatrix{4, 4, Float64, 16}: Transformation matrix [R|t] that transforms from camera to world space.

camera::Camera: Camera associated with this frame.

right_camera::Camera: In case of stereo, this is the camera associated with the right frame.

keypoints::Dict{Int64, Keypoint}: Map of that this frames observes.

ketpoints_grid::Matrix{Set{Int64}}: Grid, where each cell contains several keypoints. This is useful when want to retrieve neighbours for a certain Keypoint.

nb_occupied_cells::Int64: Number of cells in keypoints_grid that have at least one Keypoint.

cell_size::Int64: Cell size in pixels.

nb_keypoints::Int64: Total number of keypoints in the Frame.

nb_2d_keypoints::Int64: Total number of 2D keypoints in the Frame.

nb_3d_keypoints::Int64: Total number of 3D keypoints in the Frame.

nb_3d_keypoints::Int64: Total number of stereo keypoints in the Frame.

covisible_kf::OrderedDict{Int64, Int64}: Dictionary with kfid => score of ids of Frames that observe the sub-set of size score of keypoints in Frame.

local_map_ids::Set{Int64}: Set of ids of MapPoints that are not visible in this Frame, but are a potential candidates for remapping back into this Frame.

SLAM.FrontEnd — Type

Front-End component which is responsible for tracking keypoints and computing poses for the Frames. It also decides when the system needs a new Keyframe added to its map.

Parameters:

current_frame::Frame: Current frame that is being processed. This is a shared Frame between FrontEnd, MapManager, Estimator and SlamManager.

motion_model::MotionModel: Motion model that is used to predict pose for the Frame before the actual pose for it was computed.

map_manager::MapManager: Map manager that is responsible for the creation of new Keyframes in the map.

params::Params: Parameters of the system.

current_image::Matrix{Gray{Float64}}: Current image that is processed.

previous_image::Matrix{Gray{Float64}}: Previous processed image.

current_pyramid::ImageTracking.LKPyramid: Pre-computed pyramid that is used for optical flow tracking for the current_image.

previous_pyramid::ImageTracking.LKPyramid: Pre-computed pyramid that is used for optical flow tracking for the previous_image.

SLAM.Keypoint — Type

Keypoint is a feature in the image used for tracking.

Parameters:

id::Int64: Id of the Keypoint.

pixel::Point2f: Pixel coordinate in the image plane in (y, x) format.

undistorted_pixel::Point2f: In presence of distortion in camera, this is the undistorted pixel coordinates.

position::Point3f: Pre-divided (backprojected) keypoint in camera space. This is used in algorithms like 5Pt for Essential matrix calculation.

descriptor::BitVector: Descriptor of a keypoint.

is_3d::Bool: If true, then this keypoint was triangulated.

is_retracked::Bool: If true, then this keypoint was lost and re-tracked back by match_local_map! method.

is_stereo::Bool: If true, then this keypoint has left↔right correspondence for the stereo image.

right_pixel::Point2f: Pixel coordinate in the right image plane in (y, x) format.

right_undistorted_pixel::Point2f: In presence of distortion in camera, this is the undistorted right_pixel coordinates.

right_position::Point3f: Pre-divided (backprojected) right keypoint in camera space. This is used in algorithms like 5Pt for Essential matrix calculation.

SLAM.MotionModel — Type

Constant Velocity Motion Model

SLAM.MotionModel — Method

Apply Motion Model to a given wc transformation.

SLAM.add_keyframe! — Method

Copy current MapManager's Frame and add it to the KeyFrame map. Increase current keyframe id & total number of keyframes.

SLAM.backproject — Method

Transform point from 2D to 3D by dividing by K.

Arguments:

point::Point2: Point to backproject in (y, x) format.

Returns:

Backprojected Point3f in (x, y, z = 1.0) format.

SLAM.check_new_kf_required — Method

Check if we need to insert a new KeyFrame into the Map.

SLAM.check_ready_for_init! — Method

Check if there is enough average rotation compensated parallax between current Frame and previous KeyFrame.

Additionally, compute Essential matrix using 5-point Ransac algorithm to filter out outliers and check if there is enough inliers to proceed.

SLAM.compute_parallax — Method

Compute parallax in pixels between current Frame and the provided current_frame_id Frame.

Arguments:

compensate_rotation::Bool: Compensate rotation by computing relative rotation between current Frame and previous Keyframe if true. Default is true.

only_2d::Bool: Consider only 2d keypoints. Default is true.

median_parallax::Bool: Instead of the average, compute median parallax. Default is true.

SLAM.compute_pose! — Method

Compute pose of a current Frame using P3P Ransac algorithm. Pose is computed from the triangulated Keypoints (MapPoints) that are visible in this frame.

Returns:

true if the pose was successfully computed, otherwise false.

SLAM.describe — Method

Returns:

Vector of descriptors of BitVector type and vector of keypoints' coordinates of type CartesianIndex{2}.

SLAM.detect — Method

Detect keypoints in the image.

Arguments:

current_points: Vector of points, which define circular regions, where to avoid detecting new features. This can be used to decrease the amount of similar keypoints. Pass empty vector to skip this step.

σ::Real: Standard deviation for the gaussian blur. This is used to smooth out circles in the mask to reduce the chance of detecting aliased circle corners. Default value is 3. Pass 0 to skip this step.

Returns:

Keypoints in the (y, x) format.

SLAM.do_local_map_matching — Method

Given a frame and its local map of Keypoints ids (triangulated), project respective mappoints onto the frame, find surrounding keypoints (triangulated?), match surrounding keypoints with the projection. Best match is the new candidate for replacement.

SLAM.fb_tracking! — Method

Perform Forward-Backward tracking by first tracking keypoints from previous_pyramid to current_pyramid and then track resuling keypoints in reverse order. Then filter out those forward-backward keypoints that are too far away from the original keypoints to be consistent.

Arguments:

new_keypoints: Vector where to write resulting new poisitions of keypoints. Should be of the same size as keypoints vector.

max_distance::Real: Maximum distance in pixels between Forward-Backward tracked keypoints to be considered the same keypoint.

SLAM.find_best_match — Method

For a given target_mp MapPoint, find best match among surrounding keypoints.

Given target mappoint from covisibility graph, its projection onto frame and surrounding keypoints in frame for that projection, find best matching keypoint (already triangulated?) in frame.

SLAM.get_mask — Method

Create a mask from a set of points.

Each point defines a circular region that is used to avoid detecting features in that region.

Circular regions are filled with zero(T) values, while the rest of the mask is has ones(T) values.

SLAM.in_image — Method

Check if point is in the image bounds of the Camera.

Arguments:

point: Point to check. In (y, x) format.

SLAM.is_bad! — Method

Check if MapPoint is bad and update it if so.

Bad MapPoint is a 3D point that is observed by < 2 KeyFrames and not observed by current Frame (e.g. m.observed = false). If bad, set is_3d to false.

SLAM.local_bundle_adjustment! — Method

Perform Bundle-Adjustment on the new frame and its covisibility graph.

Minimize error function over all KeyFrame's extrinsic parameters in the covisibility graph and their corresponding MapPoint's positions. Afterwards, update these parameters.

SLAM.map_filtering! — Method

Filter out KeyFrames that share too many MapPoints with other KeyFrames in the covisibility graph. Since they are not informative.

SLAM.match_local_map! — Method

Try matching keypoints from frame with keypoints from frames in its covisibility graph (aka local map).

SLAM.project — Method

Project point from 3D space onto the image plane.

Arguments:

point: Point in 3D space in (x, y, z) format.

Returns:

Projected point in (y, x) format.

SLAM.project_undistort — Method

Project point onto image plane of the Camera, accounting for the distortion parameters of the camera.

Arguments:

point: 3D point to project in (x, y, z) format.

Returns:

2D floating point coordinates in (y, x) format.

SLAM.remove_mappoint! — Method

Remove MapPoint from the map given its id.

Removing a mappoint, also update covisibility scores of the observer Frames. If MapPoint is observed by the current Frame, remove its keypoint as well.

SLAM.remove_mappoint_obs! — Method

Remove KeyFrame observation from MapPoint.

SLAM.remove_obs_from_current_frame! — Method

Remove a MapPoint observation from current Frame by id.

SLAM.to_cartesian — Method

Params: x::Point2 Point to convert to CartesianIndex in (row, col) format.

SLAM.track! — Method

Given an image and time at which it was taken, track keypoints in it. After tracking, decide if the system needs a new Keyframe added to the map. If it is the first image to be tracked, then Keyframe is always needed.

Returns:

true if the system needs a new Keyframe, otherwise false.

SLAM.undistort_pdn_point — Method

Undistort point.

Arguments:

point::SVector{2}: Predivided by K & normalized point in (y, x) format.

SLAM.undistort_point — Method

Arguments:

point::SVector{2}: Point to undistort. In (y, x) format.

SLAM.update_frame_covisibility! — Method

Update MapPoints and covisible graph between KeyFrames.

SLAM.update_mappoint! — Method

Update position of a MapPoint.