IO

SLAM.ReplaySaver — Type

ReplaySaver()

ReplaySaver is used to save/load SLAM result for the visualizer. If added to the SlamManager instead of the visualizer, it accumulates poses needed for the visualization/replay and can be later used to replay the results.

Usage:

To use it, pass it to visualizer keyword argument, when creating SlamManager. Do some work, and in the end, save it to the directory.

saver = ReplaySaver()
SlamManager(params, camera; visualizer=saver)
# ...run slam...
SLAM.save(saver, "save_dir")

Then you can load it from the save directory and use it to replay results.

SLAM.load!(saver, "save_dir")

SLAM.set_frame_wc! — Method

set_frame_wc!(saver::ReplaySaver, frame_id, wc)

Add new pose to store.

SLAM.save — Function

save(saver::ReplaySaver, save_dir)

Save results into the given directory.

SLAM.load! — Function

load!(saver::ReplaySaver, save_dir)

Load results from a given directory.